+++ lib/src/models/ble_device.dart
@@ -12,7 +12,10 @@ class BleDevice {
   Uint8List? manufacturerDataHead;
   Uint8List? manufacturerData;
 
-  Future<BleConnectionState> get connectionState async => await UniversalBle.isConnected(deviceId) ? BleConnectionState.connected : BleConnectionState.disconnected;
+  Future<BleConnectionState> get connectionState async =>
+      await UniversalBle.isConnected(deviceId)
+          ? BleConnectionState.connected
+          : BleConnectionState.disconnected;
 
   BleDevice({
     required this.deviceId,